+++ src/test/jmh/nbbrd/picocsv/jmh/Benchmark.java
@@ -1,35 +1,39 @@
 /*
  * Copyright 2019 National Bank of Belgium
- * 
- * Licensed under the EUPL, Version 1.1 or - as soon they will be approved 
+ *
+ * Licensed under the EUPL, Version 1.1 or - as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
- * 
+ *
  * http://ec.europa.eu/idabc/eupl
- * 
- * Unless required by applicable law or agreed to in writing, software 
+ *
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
-package _benchmark;
+package nbbrd.picocsv.jmh;
 
 import org.openjdk.jmh.runner.Runner;
 import org.openjdk.jmh.runner.options.Options;
 import org.openjdk.jmh.runner.options.OptionsBuilder;
 
 /**
- *
  * @author Philippe Charles
  */
 public class Benchmark {
 
     public static void main(String[] args) throws Exception {
         Options options = new OptionsBuilder()
-                .include(PicocsvBenchmark.class.getSimpleName())
-                .include(FastCsvBenchmark.class.getSimpleName())
+//                .include("PicocsvBenchmark.read")
+                .include("PicocsvBenchmark.write")
+//                .include("FastCsvBenchmark.read")
+//                .include("FastCsvBenchmark.write")
+//                .forks(1)
+//                .warmupForks(1)
+//                .warmupIterations(3)
                 .build();
         new Runner(options).run();
     }